Bugfixes and Preview of Entity Manager Debugger
Bug fixes:
- issues with the hash of GroupMatcher (hashes of that struct were not unique)
- EntityManager events will now be called at ones, after OnStart, OnUpdate and OnFixedUpdate.
- Some bugs with the Systemdebugger

Added feature:
- EntityManagerDebugger (preview): shows all entities of that manager with its component. You can also filter by  components. The visual debugger will create a gameobject for each EntityManager with a DebugEntityManagerBehaviour on it.

Note: Replace EntityManager with UnityEntityManager in all your scripts. The introduction of this is necessary because of the debugger. 
If you use the EntityManager with InjectDependency, it will create a new Manager of that base type and not of the used  UnityEntityManager! So check you scripts and change it!